pyproject.toml
[tool.poetry]
name = "funding-opportunity"
version = "0.1.0"
description = "This project will develop a scrping system for global funidng opportunities"
authors = ["bella20000605 <js5081@georgetown.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
python-dotenv = "^1.0.1"
selenium = "^4.25.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-api-python-client = "^2.149.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
